:passport_control: feat(FeatureDatasetCharacteristicController): Implement role based authorization for FeatureDatasetCharacteristicController.

+++ src/main/java/io/passport/server/controller/FeatureDatasetCharacteristicController.java
@@ -3,17 +3,20 @@ package io.passport.server.controller;
 import io.passport.server.model.FeatureDatasetCharacteristicDTO;
 import io.passport.server.model.FeatureDatasetCharacteristic;
 import io.passport.server.model.FeatureDatasetCharacteristicId;
+import io.passport.server.model.Role;
 import io.passport.server.service.FeatureDatasetCharacteristicService;
+import io.passport.server.service.RoleCheckerService;
+import org.keycloak.KeycloakPrincipal;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
@@ -30,21 +33,37 @@ public class FeatureDatasetCharacteristicController {
      */
     private final FeatureDatasetCharacteristicService featureDatasetCharacteristicService;
 
+    /**
+     * Role checker service for authorization
+     */
+    private final RoleCheckerService roleCheckerService;
+
     @Autowired
-    public FeatureDatasetCharacteristicController(FeatureDatasetCharacteristicService featureDatasetCharacteristicService) {
+    public FeatureDatasetCharacteristicController(FeatureDatasetCharacteristicService featureDatasetCharacteristicService,
+                                                  RoleCheckerService roleCheckerService) {
         this.featureDatasetCharacteristicService = featureDatasetCharacteristicService;
+        this.roleCheckerService = roleCheckerService;
     }
 
     /**
      * Read all FeatureDatasetCharacteristics or filtered by datasetId and/or featureId
      * @param datasetId ID of the Dataset (optional)
      * @param featureId ID of the Feature (optional)
+     * @param principal KeycloakPrincipal object that holds access token
      * @return
      */
     @GetMapping()
     public ResponseEntity<List<FeatureDatasetCharacteristicDTO>> getFeatureDatasetCharacteristics(
             @RequestParam(required = false) Long datasetId,
-            @RequestParam(required = false) Long featureId) {
+            @RequestParam(required = false) Long featureId,
+            @AuthenticationPrincipal KeycloakPrincipal<?> principal) {
+
+        // Allowed roles for this endpoint
+        List<Role> allowedRoles = List.of(Role.DATA_ENGINEER);
+        // Check role of the user
+        if(!this.roleCheckerService.hasAnyRole(principal, allowedRoles)){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
 
         List<FeatureDatasetCharacteristic> characteristics;
 
@@ -77,11 +96,21 @@ public class FeatureDatasetCharacteristicController {
     /**
      * Create a new FeatureDatasetCharacteristic entity.
      * @param featureDatasetCharacteristicDTO the DTO containing data for the new FeatureDatasetCharacteristic with input structure
+     * @param principal KeycloakPrincipal object that holds access token
      * @return
      */
     @PostMapping()
-    public ResponseEntity<?> createFeatureDatasetCharacteristic(@RequestBody FeatureDatasetCharacteristicDTO featureDatasetCharacteristicDTO) {
+    public ResponseEntity<?> createFeatureDatasetCharacteristic(@RequestBody FeatureDatasetCharacteristicDTO featureDatasetCharacteristicDTO,
+                                                                @AuthenticationPrincipal KeycloakPrincipal<?> principal) {
         try {
+
+            // Allowed roles for this endpoint
+            List<Role> allowedRoles = List.of(Role.DATA_ENGINEER);
+            // Check role of the user
+            if(!this.roleCheckerService.hasAnyRole(principal, allowedRoles)){
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+            }
+
             FeatureDatasetCharacteristic featureDatasetCharacteristic = new FeatureDatasetCharacteristic(featureDatasetCharacteristicDTO);
             FeatureDatasetCharacteristic savedFeatureDatasetCharacteristic = this.featureDatasetCharacteristicService.saveFeatureDatasetCharacteristic(featureDatasetCharacteristic);
             return ResponseEntity.status(HttpStatus.CREATED).body(savedFeatureDatasetCharacteristic);
@@ -96,19 +125,29 @@ public class FeatureDatasetCharacteristicController {
      * @param datasetId ID of the Dataset
      * @param featureId ID of the Feature
      * @param updatedFeatureDatasetCharacteristic FeatureDatasetCharacteristic model instance with updated details.
+     * @param principal KeycloakPrincipal object that holds access token
      * @return
      */
     @PutMapping()
     public ResponseEntity<?> updateFeatureDatasetCharacteristic(
             @RequestParam Long datasetId,
             @RequestParam Long featureId,
-            @RequestBody FeatureDatasetCharacteristic updatedFeatureDatasetCharacteristic) {
+            @RequestBody FeatureDatasetCharacteristic updatedFeatureDatasetCharacteristic,
+            @AuthenticationPrincipal KeycloakPrincipal<?> principal) {
 
         FeatureDatasetCharacteristicId featureDatasetCharacteristicId = new FeatureDatasetCharacteristicId();
         featureDatasetCharacteristicId.setDatasetId(datasetId);
         featureDatasetCharacteristicId.setFeatureId(featureId);
 
         try {
+
+            // Allowed roles for this endpoint
+            List<Role> allowedRoles = List.of(Role.DATA_ENGINEER);
+            // Check role of the user
+            if(!this.roleCheckerService.hasAnyRole(principal, allowedRoles)){
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+            }
+
             Optional<FeatureDatasetCharacteristic> savedFeatureDatasetCharacteristic = this.featureDatasetCharacteristicService.updateFeatureDatasetCharacteristic(featureDatasetCharacteristicId, updatedFeatureDatasetCharacteristic);
             if (savedFeatureDatasetCharacteristic.isPresent()) {
                 return ResponseEntity.ok().body(savedFeatureDatasetCharacteristic);
@@ -125,18 +164,28 @@ public class FeatureDatasetCharacteristicController {
      * Delete by FeatureDatasetCharacteristic composite ID using query parameters.
      * @param datasetId ID of the Dataset
      * @param featureId ID of the Feature
+     * @param principal KeycloakPrincipal object that holds access token
      * @return
      */
     @DeleteMapping()
     public ResponseEntity<?> deleteFeatureDatasetCharacteristic(
             @RequestParam Long datasetId,
-            @RequestParam Long featureId) {
+            @RequestParam Long featureId,
+            @AuthenticationPrincipal KeycloakPrincipal<?> principal) {
 
         FeatureDatasetCharacteristicId featureDatasetCharacteristicId = new FeatureDatasetCharacteristicId();
         featureDatasetCharacteristicId.setDatasetId(datasetId);
         featureDatasetCharacteristicId.setFeatureId(featureId);
 
         try {
+
+            // Allowed roles for this endpoint
+            List<Role> allowedRoles = List.of(Role.DATA_ENGINEER);
+            // Check role of the user
+            if(!this.roleCheckerService.hasAnyRole(principal, allowedRoles)){
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+            }
+
             boolean isDeleted = this.featureDatasetCharacteristicService.deleteFeatureDatasetCharacteristic(featureDatasetCharacteristicId);
             if (isDeleted) {
                 return ResponseEntity.noContent().build();
